Remove a class from the hierarchy chain for case detail
This was only being used for the case details view so I've consolidated it back together
