+++ caseworker/cases/helpers/case.py
@@ -1,34 +1,12 @@
-import rules
-from dateutil.parser import parse
-from decimal import Decimal
-
-from django.shortcuts import render
 from django.urls import reverse
-from django.utils import timezone
 from django.utils.functional import cached_property
-from django.views.generic import TemplateView
 
-from caseworker.cases.helpers.licence import get_latest_licence_status
-from caseworker.queues.forms import CaseAssignmentsAllocateToMeForm
-from core.constants import CaseStatusEnum, SecurityClassifiedApprovalsType
-
-from caseworker.cases.forms.queries import CloseQueryForm
-from caseworker.cases.helpers.ecju_queries import get_ecju_queries
-from caseworker.cases.objects import Slice, Case
-from caseworker.cases.services import (
-    get_case,
-    get_user_case_queues,
-    get_case_documents,
-    get_case_additional_contacts,
-    get_activity_filters,
-)
-from caseworker.core.constants import GENERATED_DOCUMENT
-from caseworker.core.helpers import generate_activity_filters
-from caseworker.core.objects import Tab
-from caseworker.core.services import get_user_permissions, get_status_properties, get_permissible_statuses
 from lite_content.lite_internal_frontend import cases
-from lite_content.lite_internal_frontend.cases import CasePage, ApplicationPage
-from caseworker.queues.services import get_queue
+from lite_content.lite_internal_frontend.cases import CasePage
+
+from caseworker.cases.objects import Slice
+from caseworker.core.objects import Tab
+from caseworker.queues.forms import CaseAssignmentsAllocateToMeForm
 from caseworker.users.services import get_gov_user
 
 
@@ -126,138 +104,6 @@ class CaseworkerMixin:
             "allocate_to_me_form": allocate_to_me_form,
             "allocate_and_approve_form": allocate_and_approve_form,
         }
-
-
-class CaseView(CaseworkerMixin, TemplateView):
-    case_id = None
-    case: Case = None
-    queue_id = None
-    queue = None
-    permissions = None
-    tabs = None
-    slices = None
-    additional_context = {}
-
-    def is_only_on_post_circ_queue(self):
-        queue_alias = tuple(queue.get("alias") for queue in self.case.queue_details)
-        return self.is_lu_user() and queue_alias == (LU_POST_CIRC_FINALISE_QUEUE_ALIAS,)
-
-    def get_goods_summary(self):
-        goods_summary = {
-            "names": list(),
-            "cles": set(),
-            "regimes": set(),
-            "report_summaries": set(),
-            "total_value": 0,
-        }
-        for good in self.case.goods:
-            goods_summary["cles"].update(list(cle["rating"] for cle in good["control_list_entries"]))
-            goods_summary["regimes"].update(list(regime["name"] for regime in good["regime_entries"]))
-            goods_summary["names"].append(good["good"]["name"])
-            if "report_summary_subject" in good and good["report_summary_subject"]:
-                report_summary = good["report_summary_subject"]["name"]
-                if "report_summary_prefix" in good and good["report_summary_prefix"]:
-                    report_summary = f"{good['report_summary_prefix']['name']} {report_summary}"
-                goods_summary["report_summaries"].add(report_summary)
-            # support legacy report_summary field until it is removed
-            elif good.get("report_summary"):
-                goods_summary["report_summaries"].add(good["report_summary"])
-
-            goods_summary["total_value"] += Decimal(good["value"])
-        return goods_summary
-
-    def get_destination_countries(self):
-        destination_countries = set()
-        all_parties = self.case.data.get("ultimate_end_users", []) + self.case.data.get("third_parties", [])
-        if self.case.data.get("end_user"):
-            all_parties.append(self.case.data["end_user"])
-        if self.case.data.get("consignee"):
-            all_parties.append(self.case.data["consignee"])
-        for party in all_parties:
-            destination_countries.add(party["country"]["name"])
-        return destination_countries
-
-    def get_open_ecju_queries_with_forms(self, open_ecju_queries):
-        open_ecju_queries_with_forms = []
-        for open_query in open_ecju_queries:
-            open_ecju_queries_with_forms.append((open_query, CloseQueryForm(prefix=str(open_query["id"]))))
-        return open_ecju_queries_with_forms
-
-    def get_context(self):
-        if not self.tabs:
-            self.tabs = []
-        if not self.slices:
-            self.slices = []
-        open_ecju_queries, closed_ecju_queries = get_ecju_queries(self.request, self.case_id)
-        open_ecju_queries_with_forms = self.get_open_ecju_queries_with_forms(open_ecju_queries)
-        user_assigned_queues = get_user_case_queues(self.request, self.case_id)[0]
-        status_props, _ = get_status_properties(self.request, self.case.data["status"]["key"])
-        can_set_done = (
-            status_props["is_terminal"]
-            and self.case.data["status"]["key"] != CaseStatusEnum.APPLICANT_EDITING
-            and not self.is_tau_user()
-        )
-
-        context = super().get_context_data()
-        default_tab = "quick-summary"
-        current_tab = default_tab if self.kwargs["tab"] == "default" else self.kwargs["tab"]
-        show_actions_column = False
-        for licence in self.case.licences:
-            if rules.test_rule("can_licence_status_be_changed", self.request, licence):
-                show_actions_column = True
-                break
-
-        return {
-            **context,
-            "tabs": self.tabs if self.tabs else self.get_tabs(),
-            "current_tab": current_tab,
-            "slices": [Slices.SUMMARY, *self.slices],
-            "case": self.case,
-            "queue": self.queue,
-            "is_system_queue": self.queue["is_system_queue"],
-            "goods_summary": self.get_goods_summary(),
-            "destination_countries": self.get_destination_countries(),
-            "user_assigned_queues": user_assigned_queues,
-            "case_documents": get_case_documents(self.request, self.case_id)[0]["documents"],
-            "open_queries": open_ecju_queries_with_forms,
-            "closed_queries": closed_ecju_queries,
-            "additional_contacts": get_case_additional_contacts(self.request, self.case_id),
-            "permissions": self.permissions,
-            "is_tau_user": self.is_tau_user(),
-            "hide_im_done": self.is_tau_user() or self.is_only_on_post_circ_queue(),
-            "can_set_done": can_set_done
-            and (self.queue["is_system_queue"] and user_assigned_queues)
-            or not self.queue["is_system_queue"],
-            "generated_document_key": GENERATED_DOCUMENT,
-            "permissible_statuses": get_permissible_statuses(self.request, self.case),
-            "filters": generate_activity_filters(get_activity_filters(self.request, self.case_id), ApplicationPage),
-            "is_terminal": status_props["is_terminal"],
-            "security_classified_approvals_types": SecurityClassifiedApprovalsType,
-            "user": self.caseworker,
-            "show_actions_column": show_actions_column,
-            "licence_status": get_latest_licence_status(self.case),
-            **self.additional_context,
-        }
-
-    def _transform_data(self):
-        self.case.total_days_elapsed = (timezone.now() - parse(self.case.submitted_at)).days
-        if self.case.queue_details:
-            for queue_detail in self.case.queue_details:
-                queue_detail["days_on_queue_elapsed"] = (timezone.now() - parse(queue_detail["joined_queue_at"])).days
-
-    def get(self, request, *args, **kwargs):
-        self.case_id = str(kwargs["pk"])
-        self.case = get_case(request, self.case_id)
-        self.queue_id = kwargs["queue_pk"]
-        self.queue = get_queue(request, self.queue_id)
-        self.permissions = get_user_permissions(self.request)
-
-        self._transform_data()
-
-        self.tabs = self.get_tabs()
-        self.get_standard_application()
-
-        return render(request, "case/case.html", self.get_context())
 
 
 def get_case_detail_url(case_id, case_type, queue_id):
